adding lcd

+++ src/sabretracker.py
@@ -34,8 +34,8 @@ while True:
         nmea_latitude = NMEA_buff[1]                #extract latitude from GPGGA string
         nmea_longitude = NMEA_buff[3]               #extract longitude from GPGGA string
         print("NMEA Time: ", nmea_time,'\n')
-        lat = (float)(nmea_latitude)
+        lat = nmea_latitude
         lat = convert_to_degrees(lat)
-        longi = (float)(nmea_longitude)
+        longi = nmea_longitude
         longi = convert_to_degrees(longi)
         print ("NMEA Latitude:", lat,"NMEA Longitude:", longi,'\n')
